refactor: 타이머 코드 분리
